requirements/production.txt
fastapi
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
python-multipart
sqlmodel
uvicorn
